Move common methods to _base.py
